Now mostfit receieves float repayments

+++ app/controllers/data_entry/payments.rb
@@ -78,7 +78,7 @@ module DataEntry
     
     def create(payment)
       raise NotFound unless @loan = Loan.get(payment[:loan_id])
-      amounts = payment[:amount].to_i
+      amounts = payment[:amount].to_f
       receiving_staff = StaffMember.get(payment[:received_by])
       # we create payment through the loan, so subclasses of the loan can take full responsibility for it (validations and such)
       if payment[:type] == "total"
